// libs/data-access/src/lib/chats/interfaces/chats.interface.ts
import { Profile } from '../../profile';

export interface Chat {
  id: number;
  userFirst: Profile;
  userSecond: Profile;
  messages: Message[];
  companion?: Profile;
}

export interface Message {
  id: number;
  userFromId: number;
  personalChatId: number;
  text: string | null;
  createdAt: string;
  isRead: boolean;
  UpdatedAt?: string;
  user?: Profile;
  isMine?: boolean;
}

export interface lastMessageRes {
  id: number;
  userFrom: Profile;
  message: string;
}